+++ src/main/java/com/github/markusbernhardt/selenium2library/keywords/Selenium2LibraryEnhancement.java
@@ -3,6 +3,7 @@ package com.github.markusbernhardt.selenium2library.keywords;
 import java.util.List;
 
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.remote.RemoteWebDriver;
 
 import com.github.markusbernhardt.selenium2library.Selenium2LibraryNonFatalException;
 import com.github.markusbernhardt.selenium2library.locators.ElementFinder;
@@ -13,6 +14,21 @@ public class Selenium2LibraryEnhancement extends Waiting {
 	// Keywords
 	// ##############################
 
+	public String getRemoteSessionId() {
+		try {
+			return ((RemoteWebDriver) webDriverCache.getCurrent())
+					.getSessionId().toString();
+		} catch (ClassCastException e) {
+			return "No remote session id";
+		}
+	}
+
+	public String logRemoteSessionId() {
+		String actual = getRemoteSessionId();
+		info(actual);
+		return actual;
+	}
+	
 	public void addLocationStrategy(String strategyName,
 			String functionDefinition) {
 		addLocationStrategy(strategyName, functionDefinition, null);